requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
chromedriver-binary==105.0.5195.52.0
chromedriver-binary-auto==0.2.0
click==8.1.3
Flask==2.2.2
frozenlist==1.3.1
fs==2.4.16
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
zipp==3.8.1
